Use the current working directory's config file.
Instead of loading the configuration using a relative path from
core/, load it from the current working directory.

This means that a Bedrock command should be called necessarily from a
project root directory (where the config is supposed to live), but I
guess this was already the case when using commands defined in the
Gulpfile or in the package.json file.

In other words, this change shouldn't have an impact on existing
installs, but paves the road to allow calling Bedrock from projects that
don't have the full source code of Bedrock within their own directory,
and provide their own configuration files.

See https://github.com/usebedrock/bedrock/issues/320 for details.